move remove_superficial_options into utils
This move makes sense if one wants to reuse remove_superficial_options
since it can be not specific to cache.py only.

This prepares prosopopee for Pillow support.
