+++ prosopopee/utils.py
@@ -10,6 +10,20 @@ from path import Path
 import ruamel.yaml as yaml
 
 
+def remove_superficial_options(options):
+    cleaned_options = options.copy()
+    del cleaned_options["name"]
+    if "text" in cleaned_options:
+        del cleaned_options["text"]
+    if "type" in cleaned_options:
+        del cleaned_options["type"]
+    if "size" in cleaned_options:
+        del cleaned_options["size"]
+    if "float" in cleaned_options:
+        del cleaned_options["float"]
+    return cleaned_options
+
+
 class CustomFormatter(logging.Formatter):
     """Logging Formatter to add colors"""
